analytics/streamlit/requirements.txt
streamlit==1.26.0
web3==6.9.0
python-dotenv==1.0.0